requirements/tests.txt
coverage==4.4.2
isort==4.2.15
pycodestyle==2.3.1
pydocstyle==2.1.1
pylint==1.8.1
pytest==3.3.2
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-pylint==0.7.1
